requirements.txt
fastapi
uvicorn
transformers
torch
sentence-transformers
faiss-cpu
spacy
PyPDF2
python-docx
pydantic
python-multipart
pytesseract
pdf2image

